requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
geopandas==1.1.1
fiona==1.10.1
shapely==2.1.1
folium==0.20.0
matplotlib==3.10.6
reportlab==4.4.4
weasyprint==66.0
pillow==11.3.0
requests==2.32.5
selenium==4.35.0
webdriver-manager==4.0.2
pandas==2.3.2

